+++ Ball.hpp
@@ -9,16 +9,16 @@ class Ball {
 private:
 	int R = 0, G = 0, B = 0;
 	int opaque = 255; // 255 = fully opaque, closer to 0 = more transparent
-	int posx, posy;
+	float posx, posy;
 	int width;
 public:
-	Ball(int StartX, int StartY, int StartWidth) : // creates paddle shape
+	Ball(float StartX, float StartY, int StartWidth) : // creates paddle shape
 		posx(StartX), posy(StartY), width(StartWidth) {} // with position
 
 	void SetColour(int R, int G, int B);
 	void RenderBall(SDL_Renderer* renderer);
 
-	void Move(int changex, int changey);
+	void Move(float changex, float changey);
 
 	int GetX();
 	int GetY();
